+++ apps/nestjs-indexer/src/indexer/buy-token.indexer.ts
@@ -6,6 +6,7 @@ import { validateAndParseAddress, hash } from 'starknet';
 import { BuyTokenService } from 'src/services/buy-token/buy-token.service';
 import { IndexerService } from './indexer.service';
 import { ContractAddress } from 'src/common/types';
+import { safeUint256ToBN } from './utils';
 
 @Injectable()
 export class BuyTokenIndexer {
@@ -56,37 +57,6 @@ export class BuyTokenIndexer {
     }
   }
 
-  private safeUint256ToBN(
-    lowFelt: starknet.IFieldElement,
-    highFelt: starknet.IFieldElement,
-  ): bigint {
-    try {
-      // Convert FieldElements to BigInts directly
-      const low = FieldElement.toBigInt(lowFelt);
-      const high = FieldElement.toBigInt(highFelt);
-
-      this.logger.debug(`Converting uint256 - low: ${low}, high: ${high}`);
-
-      // Validate the low and high values
-      const UINT_128_MAX = BigInt('0xffffffffffffffffffffffffffffffff');
-      if (low > UINT_128_MAX || high > UINT_128_MAX) {
-        this.logger.warn(`Low or high value exceeds maximum ${UINT_128_MAX}`);
-        // Handle overflow by capping at max value
-        return UINT_128_MAX;
-      }
-
-      // Combine high and low parts into a single bigint
-      const fullValue = (high << BigInt(128)) + low;
-
-      this.logger.debug(`Full value: ${fullValue}`);
-
-      return fullValue;
-    } catch (error) {
-      this.logger.error('Error converting uint256:', error);
-      return BigInt(0);
-    }
-  }
-
   private async handleBuyTokenEvent(
     header: starknet.IBlockHeader,
     event: starknet.IEvent,
@@ -134,31 +104,25 @@ export class BuyTokenIndexer {
         quoteAmountHigh,
       ] = event.data;
 
-      const amountRaw = this.safeUint256ToBN(amountLow, amountHigh);
+      const amountRaw = safeUint256ToBN(amountLow, amountHigh);
       const amount = formatUnits(amountRaw, constants.DECIMALS).toString();
 
-      const priceRaw = this.safeUint256ToBN(priceLow, priceHigh);
+      const priceRaw = safeUint256ToBN(priceLow, priceHigh);
       const price = formatUnits(priceRaw, constants.DECIMALS);
 
-      const protocolFeeRaw = this.safeUint256ToBN(
-        protocolFeeLow,
-        protocolFeeHigh,
-      );
+      const protocolFeeRaw = safeUint256ToBN(protocolFeeLow, protocolFeeHigh);
       const protocolFee = formatUnits(
         protocolFeeRaw,
         constants.DECIMALS,
       ).toString();
 
-      const lastPriceRaw = this.safeUint256ToBN(lastPriceLow, lastPriceHigh);
+      const lastPriceRaw = safeUint256ToBN(lastPriceLow, lastPriceHigh);
       const lastPrice = formatUnits(
         lastPriceRaw,
         constants.DECIMALS,
       ).toString();
 
-      const quoteAmountRaw = this.safeUint256ToBN(
-        quoteAmountLow,
-        quoteAmountHigh,
-      );
+      const quoteAmountRaw = safeUint256ToBN(quoteAmountLow, quoteAmountHigh);
       const quoteAmount = formatUnits(
         quoteAmountRaw,
         constants.DECIMALS,
